feat: Decouple creating MavenSettings from ExecutionContext
Creating/parsing a MavenSettings currently emits a side effect of
sending a bunch of information to ExecutionContext. This change splits
this concept into two methods, one to parse/generate the MavenSettings,
and one to send the information to the execution context.

BREAKING CHANGE: Anything that uses `MavenSettings#parse` will break following this
change and need to instead generate the settings and then call
#sendToExecutionContext