Change how indentation is autodetected
For each indent value detected, we count all the lines whose indent is a
multiple of the current one.
This way, each potential candidate is evaluated together with the other
lines.

Also we don't count spaces and tabs in the same way and we don't count
lines that mix both formats.