Fix parent/module resolution in MavenParser
Creating a copy is problematic as we are then pointing to stale objects
that do not contain the whole updated information.
In the case of the test, the project in the middle of the hierarchy
loses its parent in the process.
Then when maybeUpdateModel() is called, the pom file cannot be
downloaded and an exception is thrown, resulting in the document not
being updated.
This uses the same pattern for modules as what is currently used to set
the parent.

Related to the issues described in #1582.