Make `HiddenField` recipe predictable (#2596)
By using a `LinkedHashSet` rather than a regular `HashSet` in `FindNameShadows#find()` to collect the shadowing variables, the renaming of the variables becomes predictable. I.e. the same variable always gets renamed to the same name if the recipe is tested multiple times. This is especially important for tests.

Fixes #2592