Slightly optimizes the `JavaTypeCache`
The Snappy compression used in `JavaTypeCache` is typically not able to compress signatures which are shorter than about 50 characters. The compressed representation is often even one or two bytes larger.

Thus this commit aims at slightly optimizing the cache by using the signature String object itself as the cache key in these cases. As a result the cache keys are then around 6 bytes longer (when accounting for String's 24 byte object size vs. BytesKey's 16 bytes), but the whole parsing phase is about 5% faster.

Different thresholds for when to start compressing could make sense in different situations, so maybe this ought to be configurable.