Fix issue with configuring rewrite-java-* Gradle projects (#2633)
This initially presented only on Windows, but could easily have presented on any other operating system. Essentially, the `compatibilityTest` testing task was reaching into another project's configuration and specifically trying to get informaton about it's `sourceSet` configuration settings. However, if that project was not yet configured, then it would not have had the `JvmEcosystemPlugin` applied to it. Without the `JvmEcosystemPlugin`, this would mean that there was no `sourceSet` container to get information from ultimately leading to one of the rewrite-java-* projects failing to configure with a rather vague error messaging stating these same qualities. At first glance, it would appear that the `sourceSet` container was missing from the current project (ie. rewrite-java-*), while in fact it was missing from the rewrite-java-tck project (which hadn't yet been configured).