Fix MavenArtifactDownloader - Closed InputStream
MavenArtifactDownloader was automatically closing the input stream
generated when an artifact is downloaded because the response was
autocloseable and is instantiated in a try statement.

This fix uses the method #getBodyAsBytes() instead and creates
a ByteArrayInputStream. This means that the jar file is stored
on memory, which it should not be an issue.