New `InstanceOfPatternMatch` recipe
Implement a recipe which for Java 17+ transforms `instanceof` expressions which have a corresponding type cast within their flow scope (see https://openjdk.org/jeps/394 for details) to an `instanceof` with a pattern variable and then also substitutes the type cast expressions accordingly.

Currently, this works inside any binary boolean expressions and the then-block of `if` statements and ternary operator expressions.