`FinalizePrivateFields` also needs to update the `JavaType`
Variable declaration modifiers are both stored as `J.Modifier` objects in the LST and as flags on the corresponding `JavaType.Variable` object. The recipe was only updating the former, which had the consequence that the recipe didn't play nicely with for example `RenamePrivateFieldsToCamelCase`, which is based on the latter.

Related to: #2611