Improve the equality check in `ListUtils#flatMap()`
In case the mapping function returned an `Iterable` for one or more elements, the `flatMap()` method will now return the input list if the elements are all identical.

Also, in `map()` the logic for deleting `null` elements is slightly improved by only iterating over the list if it contains `null`s.