Fix the replacement coordinates for unary expressions (#2784)
A unary expression (like e.g. `!b`) cannot be used as a statement on its own like modifying unary expressions (like `i++`) can. As a result the `JavaTemplate` must use a corresponding coordinate when replacing a unary expression, as otherwise the Java code produced by the template compiler is incorrect.