`MethodParamPad` no longer formats entire method body
The `MethodParamPad` recipe now uses the first method parameter (or corresponding `J.Empty` marker if there are none) as the stop criterion for the `SpacesVisitor` rather than the method declaration itself.

To get this to work in the context of the `MethodParamPad` the `SpacesVisitor` had to be changed to add the stop message to the root cursor, since when visiting a subtree there is not necessarily any cursor for the source file node.