Fix `IsEmptyCallOnCollections` in for loop control
Uses `JavaTemplate` to replace the `J.Binary` rather than the `J.MethodInvocation` as that is also what is the intention.

This required to change the `JavaTemplate` engine to also use the `BlockStatementTemplateGenerator` when replacing expressions, since the template may refer to declarations visible to the insertion point in the target class.

Fixes #2813