First draft of JSON schema for OpenRewrite resources
This is a first draft of the JSON schema for OpenRewrite resource descriptors. Although the resource descriptor types themselves currently include a version "v1beta", this schema currently doesn't have any version number.

The schema applies to all resource types and uses the `type` property as a discriminator to declare the additional properties that apply to the respective resource type.

Whenever a new resource type is added or a resource type is modified, this schema must also be modified accordingly.

Fixes #2823