Update the CodeQL data flow models
Updates the `model.csv` and `sinks.csv` files using the most current CodeQL release 2.12.2.

The queries had to be slightly changed, but the output is basically the same (except for the `generated` column which has been renamed to `provenance` with the values `manual` and `generated`). Also, `codeql query run` outputs BQRS files which can then be transformed to CSV using `codeql bqrs decode`. The resulting CSV files look slightly different (with `,` as separator rather than `;` and strings being quoted).

Please note that the new models contain more data and are presumably more accurate.