Improve performance of `MethodMatcher`
Before the `MethodMatcher` uses regular expressions to match the type name, method name, and arguments. This has quite a high overhead compared to using `equals()`, which in many cases is applicable.

The code is messier now, but should provide substantially better performance. The code will be cleaned up in subsequent commits.