GitProvenance Support for non build environments without .git config
Currently, GitProvenance is mainly calculated using the information that is
inside the .git config. Only there is a special behavior when Jenkins is
the BuildEnvironment.

However, there is no .git config in GitLab, GitHub and Drone when
a repository is checkout and the required Git metadata is on
environment variables.

Therefore, this patch adds BuildEnvironment#buildGitProvenance() to
force an implementation for CI BuildEnvironments. If they need
.git config information, such as Jenkins or Travis, they throw
IncompleteGitConfigException.

A new CustomBuildEnvironment class has also been added to support
the use case of fetching external repositories from a GitHub repository.