Support `instanceof` pattern variables in `JavaTemplate` (#2863)
Using `instanceof` pattern matching an `if` statement can declare local variables which can be dereferenced in the condition itself, in the `then` block, the `else` block, or also in successor statements to the `if` statement. Therefore, it is important that `JavaTemplate` keeps these `instanceof` pattern matching expressions when generating the stub to extract the type attributed substitution of a template.

For any `if` statements which contain `instanceof` pattern variables the `if` statement is added in a reduced form to the stub. The flow scoping rules of the pattern variables require the then and else blocks to retain the same "completes normally" characteristics as in the original code, as this influences the scope of the pattern variable.

Additionally, this PR fixes a few other minor problems with the stub generation in the `JavaTemplate` engine. Specifically, any variables declared in the same block as the element to be replaced, but preceding it, are now also generated into the stub. This was also the case before, but didn't work in the context of loops or labeled blocks.

Finally, `J.Ternary` expressions with pattern variables in the condition are also supported.

Fixes: #2820