Allow matching an LST subtree against a `JavaTemplate`
`JavaTemplate` offers a new `matches(J)` method which can be used to check if the given tree matches the receiving `JavaTemplate`. This uses `SemanticallyEqual#matchesTemplate()` and will check if the tree at the cursor is identical to the tree represented by the template.

The template can also contain `#{any(...)}` parameters, which require the input subtree to contain a typed subtree at the given location with a type compatible to the type of the template parameter.