Use `@ExpectedToFail` annotation
Using JUnit Pioneer's `@ExpectedToFail` rather than `@Disabled` has the nice benefit that if someone fixes the issue unknowingly (e.g. as a side effect of fixing some other issue), the annotated test case will show up as a failure and the developer will be forced to consider resolving the corresponding issue as well.