Add a new `@Primitive` annotation for `JavaTemplate#compile()` (#2959)
When using `JavaTemplate#compile()` to build a template based on a Java lambda, there needs to be a way to create templates which refer to primitive template parameters.

Since the SAM types (`F0` through `F10`) don't allow for this, instead add an `@Primitive` annotation type with source retention. Any parameter annotated with this on the lambda will in the generated template string use the corresponding primitive type instead (e.g. `#{any(int)}`).