Add `JavaTemplate.Matcher`
In order for a visitor to be able to extract the subtrees matched by the placeholders in the `JavaTemplate` pattern, there is now a new `JavaTemplate.Matcher` class along with a new `JavaTemplate#matcher()` method.

Using this `Matcher` the client code can call `find()` and subsequently `parameter(int)` to extract the matched parameters.