Fix `JavaTemplate` for `J.Try` with resources
The resources of a try-with-resources statement should only be generated by `BlockStatementTemplateGenerator` if generating code for a statement within the try-block, as the resources are not visible in the catch or finally blocks.

The case when generating code for one of the actual resources is not 100% correct yet, as a resource may refer to prior resources.