Handle static imports in `SemanticallyEqual`
A reference to a statically imported field (or also a locally declared field or variable) is represented as a `J.Identifier` rather than a `J.FieldAccess` element in the LST. As a consequence `SemanticallyEqual` now compares the type-attribution on those, so that for example a fully qualified reference to a static field is regarded as identical to an unqualified reference with a corresponding static import.

`JavaTemplateJavaExtension` had to be modified for the `J.FieldAccess` case, because a field access should be treated as an expression rather than a statement.