Allow `JavaTemplate` to use shared `JavaTypeCache`
As currently every application of `JavaTemplate` leads to a new `JavaParser` getting created with its own `JavaTypeCache` the result is that every time a `JavaTemplate` get applied, the LST will end up containing type attribution linked to identical yet distinct copies of `JavaType` objects, even if the LST elements refer to the same Java elements on the classpath. This in turn can cause a lot of memory overhead.

This commit aims at mitigating this issue by allowing `JavaTemplate` (specifically `JavaTemplateParser`) to use a shared `JavaTypeCache`. The life cycle of the `JavaTypeCache` is the same as that of the root `Cursor`, which typically gets instantiated by the `RecipeScheduler`, and it will get used by every `JavaTemplate` with the exact same classpath. At the end of each compilation performed by the `JavaTemplateParser` the `JavaType` objects which were created for the parsed Java sources get purged again.

Since the `JavaTypeCache` structure itself isn't thread-safe, each `JavaTemplateParser` will create a clone of the shared `JavaTypeCache` and work with that.