Optimize `StringUtils#matchesGlob()`
This Glob matching utility was using Ant's `SelectorUtils#match()` method, which uses `String#toCharArray()`. Often it is however faster to work with `String#charAt()`. Performance measurements indicate that the implementation is now more than twice as fast and since it is used very intensively in `rewrite-maven` (to match GAVs), the boost will be visible in some recipes.