Allow upgrading non-semver Maven dependencies
When an existing Maven dependency has a version which is non-semver-friendly (like `org.springframework.cloud:spring-cloud-dependencies:Camden.SR5`), the `UpgradeDependencyVersion` and `ChangeParentPom` recipes should be able to do the upgrade anyway.

Introduces a new `Semver#isVersion(String)` utility to check if a given version string is a valid semver version.