Reduce memory overhead of `JavaType` implementations
When Jackson deserializes an array into a `List` it by default creates an `ArrayList` with the default capacity (10) and populates it with the values. For `JavaType.Method` that is typically quite wasteful, since it uses lists for parameter names, parameter types, thrown exceptions, and annotations and these lists are very often shorter.

One alternative is to register a custom `ValueInstantiator` for `ArrayList` which creates lists with a smaller capacity (e.g. 4 or even just 1).

Another alternative, as done in this commit, is to change the internal representation to use arrays and create a corresponding `@JsonCreator` annotated constructor.