`UsesMethod` should rely on `TypesInUse`
If `UsesMethod` does not find any reference to the desired method in `TypesInUse` then it should return rather than traversing the entire LST in an attempt to find a match.

This change is being made for performance reasons, because this visitor is used very frequently as a precondition by recipes.