Improve OrderImports description
When I originally looked at this recipe, I was a bit confused on
what it does as the description was very generic. I think that
providing more information would help people who view this recipe
in our docs or in the saas who won't necessarily take the time to
dig into the code.