Ensure that even system-installed artifacts are resolved.
When running Maven on Linux from a system maintainer-installed package
and adding a dependency to a recipe that is also a dependency of Maven
itself (e.g. slf4j-api), then the system-installed version is used (e.g.
file:/usr/share/java/slf4j-api.jar). Unfortunately, this path is not
matched be the regular expression, since no version is added to the file
name. This results in an incomprehensible error "Unable to find runtime
dependencies beginning with: slf4j-api" that gives no clue what's wrong.
Therefore, its better to also output the current class path being
scanned.