Add FindMethods MethodCalls columns for type and method
Particularly helps cases with wildcards, as `*..*Utils *(..)` does not always produce a data table row that can be traced back without context.
Having the type and method name available, and split out from additional context, makes it easier to sort & filter on values as opposed to the usage line only.