Correctly handle empty parens before trailing lambda
A function which takes a single lambda argument can be called either with parentheses around the lambda, without parentheses, or with empty parenthesis followed by the lambda. This last calling style was not yet handled correctly by the parser.

To do this correctly we should probably add another marker to capture the whitespace which could be inbetween the opening and closing parenthesis.

Also, the logic in `mapFunctionalCallArguments()` is still quite convoluted and can be simplified even more.