Delay the resolution of classpath in `JavaParser.Builder` (#3505)
Instead of directly resolving artifact names to classpath entries inside `JavaParser.Builder#classpath(String...)` this is now delayed until `build()` is called for the first time. This has the advantage that constructing a builder is cheap even when the `classpath(String...)` method gets invoked. As a result a recipe can initialize a `JavaTemplate` (with a customized `JavaBuilder`) in a field initializer of its visitor without incurring a rather big overhead.