Standardize line break recipe name
Right now we have a recipe with the display name of `Normalize the line breaks` and a recipe with the display name of `Normalize line breaks` (one for Java and one for XML). Let's make them both say the same thing since they do the same thing in their respective language.