Make parsing and printing of arguments more consistent
Fixing the parsing of trailing constructor lambdas turned out to require quite a big refactoring of how argument lists are parsed and printed. But I feel that it is more consistent and more robust now, since the compiler's AST is now used to determine if a call has a trailing lambda or not, whereas before this was "parsed".

Fixes: #278