Fix parsing of calls to implicit extension functions
An extension function can also be "implicitly" declared when it is declared by a corresponding function type. These extensions are then called in the same way as explicit extension functions (i.e. with the receiver before the dot). This is now handled properly by the parser, although it has to query the AST to determine if this is the case.

Fixes: #270