Polish `TypeUtils` API
To reduce the API surface area, the `fullyQualifiedNamesAreEqualAsPredicate()` method was removed again (only one usage). Also, `getFullyQualifiedClassPath()` was renamed to `toFullyQualifiedName()`.

Note: According to the JLS this would be called the _canonical name_, whereas the name with dollars in it would be the _binary name_. But for consistency I stayed with the _fully qualified name_ nomenclature.