Lazily evaluate chained validations
To support lazy evaluation of
someValidation.and(Validate.test(predicate)), the predicate MUST NOT be
evaluated in the test method already. Rather return an anonymous
subclass that only evaluates the predicate in isValid().

To still support the reporting of validation failures, the anonymous
subclass must be of type Invalid, even though the validation result may
be valid.

Out of the 3 new unit tests, the second one fails without this change,
showcasing the original problem.