Fix formatting of calls without parentheses
In a call without parentheses (as for example in `it.name to it.get(this)` where `to` is the method) the trailing space should not be used to right-pad the last argument. Instead, that should belong as left padding of the following element. This is important to be able to correctly implement formatting.