Improve performance of `UsesType`
`UsesType` is extensively used as a predicate by recipes and it is thus important to be efficient.

Typically, `UsesType` is used in one of three ways:
 1. To do an exact match on a type
 2. To match all types in a given package
 3. To match all types in a given package or subpackages

For 1. the code was already optimized in the past (see https://github.com/openrewrite/rewrite/commit/071185a5b10a207bff46c69d16c18db695be0f0d). For 2 and 3 the code currently creates a regex pattern and uses that for the matching. Both the pattern creation and matching can be expensive. Thus, this PR optimizes these two cases by doing a simple prefix match.