Add missing recurring dependencies to `ResolvedDependency`
When `ResolvedPom#resolveDependencies()` processes the dependencies, it skips any dependencies it has already seen. But it must still make sure to add them to the list of dependencies of the artifact that requested it.

Also exclude `ResolvedDependency#dependencies` from `hashCode()` since dependencies can be circular (although not within one scope).

Fixes: #3626