Use a `Set` for `RecipeScheduler.RecipeRunCycle#madeChangesInThisCycle`
Use a `Set` (specifically a `Set` based on `IdentityHashMap`) to track the recipes that made changes in a cycle. This improves performance a bit, because the recipe instances are typically the same and then there is no need to query the same recipe repeatedly for `causesAnotherCycle()`.

Also attempt to revert 0902de56473f03f11bf1b2262a0039bec16f40f8 so that the recipe scheduler when possible stops after the first cycle.