Change UpdateGradleWrapper to prefer to keep the existing distributionUrl.
Still uses services.gradle.org to query for available Gradle distributions, so it is possible for dynamic version selectors to fill in a wrapper *not* available on the internal repository. This is because Gradle distributions are typically hand-uploaded to bespoke places within internal artifact repositories, so there is no consistent way to query their availability.