Update `AnnotationService` API to use `Cursor`
The `AnnotationService` API should operate on `Cursor` rather than `J` instances. The reason being that in the case of Kotlin a use-site targeted annotation could be declared at a higher level than the element itself. An example would be something like `@getter:Suppress` on a property with an explicit getter.