Constructor call type attribution should include generics
When the target type of a constructor call declares generic type variables, the `J.NewClass` type attribution should also include these details. So for example the type attribution for `new ArrayList<String>()` should be a `JavaType.Method` which as the declaring type has `java.util.ArrayList<String>` rather than just `java.util.ArrayList`.

An alternative to this would be to add a new `type` field to `J.NewClass`.