Make `MavenParser` stricter (#3937)
Since a `ResolvedGroupArtifactVersion` is expected to have non-`null` values for `groupId` and `version`, the `RawPom#toPom()` method now also enforces this and throws an exception when `null` values are present. This in turn makes the parser as a whole more robust as it will immediately convert this to a `ParseError` rather than failing hard later on.