Fix `AddDependency` when dependent on other recipes
The `AddDependency#onlyIfUsing` option currently only works if the source files already in the initial state (before any other recipes were applied) already contained references to what is specified in the `onlyIfUsing` option. Thus, if another recipe were to modify the code such that it adds a reference to code matching `onlyIfUsing`, this edit on its own would not lead to the dependency being added.

The problem is that `AddDependency` only runs in one cycle and thus uses the scanning phase to decide if a dependency needs to be added. To properly describe the use case described above, it should however request a second cycle in some scenarios.