Convert `FindCommitters` to standard recipe (#3991)
* Convert `FindCommitters` to standard recipe

As all source files in a repo contain the same (or identical) `GitProvenance` marker, there is no need for a scanning recipe. Instead, the data tables can be fully generated from the first source file which has a `GitProvenance` marker.

The recipe uses a transient field to remember if the data table has been generated already. An alternative would be to use a message in the `ExecutionContext`.

* Override `maxCycles()` for clarity