`MethodMatcher`: Remove `Class.forName()` call
In the CLI this causes a performance problem for some reason. But since the documentation of `MethodMatcher` states that the names need to be fully qualified, I think it is best if we completely remove the `Class.forName()` call.

As we probably have at least some usages with `Integer`, `Object`, etc. I think it is best if we at least support the most common of these. If we want to support them all, we could also enumerate them all, which would be better than calling `Class.forName()`.

Problematic are common `MethodMatcher`s like `new MethodMatcher("DependencyHandlerSpec *(..)")` as we have them in e.g. `ChangeDependencyArtifactId`.