Make `JavaTemplate` engine extensible (#3475)
* Add new `JvmParser` as supertype to `JavaParser`

To allow the `JavaTemplate` mechanism to be used for other languages like Kotlin, the provided parser builder must provide a way to add a classpath entry. This is because the internal `__M__` and `__P__` types are required for the parameter substitution and rather than including them as source code in the compilation unit, the parser will load them from the classpath.

To this end this PR introduces new `JvmParser` and `JvmParser.Builder` types and the `JavaParser.Builder#classpath()` methods are moved to `JvmParser.Builder`. In order to not increase the API surface area, no corresponding getter was added to `JvmParser.Builder`. Instead, there is a new `Internals` class providing a static `getClasspath(JvmParser.Builder)` method.

Issue: openrewrite/rewrite-kotlin#218

* Correct bounds on `JvmParser.Builder`

* Make `JvmParser.Builder` constructor protected

* No need for initializer block when templating context-free expression

* Expose more of `JavaTemplate` engine API for extension

* Remove `JvmParser` and `Internals` again

* Make `Substitutions` extensible

* Lazily initialize `TEMPLATE_CLASSPATH_DIR` at runtime

* Add `BlockStatementTemplateGenerator#TEMPLATE_INTERNAL_IMPORTS`

* Correction to `Substitutions` after merging

* Rename new `addClasspath()` to `classpathEntry()`

* Remove unused `JavaTemplate#parameterCount`

* Update rewrite-java/src/main/java/org/openrewrite/java/internal/template/__M__.java

* Update rewrite-java/src/main/java/org/openrewrite/java/internal/template/__P__.java

* Revert accidental visibility change

* Fix problem with classpath in `JavaParser.Builder`