Correct UTF-8 BOM detection in `EncodingDetectingInputStream` (#4139)
The sequence `C3 AF C2` is not actually the UTF-8 BOM sequence. The BOM sequence is the `\uFEFF` character, which in UTF-8 is encoded as `EF BB BF`. But if these three bytes were to get encoded as individual characters using UTF-8, then the sequence `C3 AF C2` is the first half of `C3 AF C2 BB C2 BF`. But note that this is not a proper UTF-8 BOM, but a BOM that erroneously got encoded again.