Default tests to supplying the runtime classpath to the default JavaParser.
We're not entirely sure why this wasn't already the default. Jonathan and I think it may be because at one point the JavaParser was responsible for constructing a JavaSourceSet marker listing the runtime classpath, and that was an expense to be avoided. But that is no longer the responsibility of JavaParser.