Rename `J.InstanceOf.Padding#getExpr()` and `J.LambdaParameters.Padding#getParams()` (#4163)
* Rename `J.InstanceOf.Padding#getExpr()`

The name should be in line with `J.InstanceOf#getExpression()` and the `J.InstanceOf#expression` field, therefore renaming. Keeping the old method as deprecated for backward compatibility.

* Rename `J.LambdaParameters.Padding#getParams()`

The name should be in line with `J.LambdaParameters#getParameters()` and the `J.LambdaParameters#parameters` field, therefore renaming. Keeping the old method as deprecated for backward compatibility.