YAML: Support lists of lists
YAML supports lists of lists. They are not common but sometimes
required, for example, in some AWS CloudFormation templates.
Currently, the YamlParser doesn't correctly parse these lists. It adds
additional hyphens.

Original     ->     After parsing
root:               root:
  - - a               - - - a
    - b                 - b

This happens for every line that contains multiple hyphens (like for the
'a' line in the example above).
That's because the parser interprets these hyphens twice.
Once as a SequenceEntry and once as prefix for the last block of that
line (in this example, the scalar 'a').

The bug is in SequenceBuilder.push. This method builds the prefix of a
given block by using the substring between the FIRST hyphen and the
block itself. So if there is more than one hyphen in that row, the
additional hyphens will be included in the prefix, although they are
already represented in the LST as SequenceEntries.

Fix it by using the substring between the LAST hyphen and the block
itself. To do this, extend the functionality of the commentAwareIndexOf
method, so that it can find the first or the last index. Don't create
an additional method as not to duplicate the logic. Also use a dedicated
enum because it is cleaner and much more expressive than using a boolean
flag.

Fixes #4176