Fix JavaParser.dependenciesFromResources to always return resources even when there are already matching entries in ~/.rewrite/classpath.
This fixes a scenario where the function could return different values based on the pre-existing contents of that directory.

Before this change dependenciesFromResources() preferred a jar that is already extracted into ~/.rewrite/classpath over classpath resources. But since it's a prefix match the same prefix might match a different jar in ~/.rewrite/classpath than it does in the jar.
For example a jar contains "snakeyaml-2.2.jar" and ~/.rewrite/classpath contains "snakeyaml-1.3.jar" the dependenciesFromResources() when called with the query "snakeyaml" will use 1.3, even though the author of the recipe/module packed in 2.2 and expects that to be used