Check for annotations on qualified type
The `AnnotationService` will now also return any type use annotation that is on the type of a variable is in `J.@Nullable Identifier label`.

At some point the `AnnotationService` should probably offer more precise methods, which allow checking for annotations with a given target (e.g. type use), so that the expected annotations are returned / matched.