Handle case-insensitive document declaration in the XML parser
Current we are strict on the document declaration for files which are parsed by the XML parser, the document declaration must be in upper case. Going forward support document declarations regardless of their case.

We want to do this to allow the XML to parse HTML files, which aren't as strict in their formatting