Check that we have downloaded the file successfully
When we are downloading files we should check that the http call was successful instead of assuming that because no exception was thrown that the http status was successful. In the case were we get a 408, we don't throw an exception. This results in a InputStream of null to be returned. When reading from a InputStream of null, it infinitely returns a list of null bytes. In our use case we end up writing null to a temporary file until we run out of disk space

We can also make this change in the `HttpSender` and throw when we get a none successful http code