Override Lombok handler for `@SneakyThrows`
Adds code to allow us to override the default handler for the `@SneakyThrows` annotation. This is necessary because the handler for this annotation doesn't only add generated declarations, but it also modifies the input source code by adding an outermost try-catch block. This in turn breaks OpenRewrite, because we no longer have access to AST corresponding to the original source text.