Allow multiple search results to stack up on the same element.
The "one search result per element" was a choice we made for aesthetic reasons as much as anything. Since then we've found that sometimes there is reason to run a sequence of search recipes,checking between each one to see if it made changes. With the old logic this procedure could result in an inaccurate idea about which of the search recipes had found what they were looking for. e.g.: running HasMinimumJavaVersion in sequence